feat(nav): move tabs into authenticated layout
