+++ src/renderer/root/components/AuthenticatedLayout/Tab/Tab.js
