+++ src/renderer/root/components/AuthenticatedLayout/Tab/index.js
