+++ src/renderer/root/components/AuthenticatedLayout/Tabs/Tabs.scss
